refactor(active-walk): extract stats and controls into WalkStatsSheet
Stage 1 of collapsible stats panel redesign. Pure extraction — no
visual or behavioral change. Pulls statsSection, controlsSection,
micButton, and actionButton out of ActiveWalkView into a new
WalkStatsSheet component.

Later stages will add collapsible state, drag gesture, and full-screen
map background. This extraction makes those changes localized.
